% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banknote_help.R
\docType{data}
\name{banknote}
\alias{banknote}
\title{Banknote dataset}
\format{An object of class \code{data.frame} with 200 rows and 7 columns.}
\source{
\href{http://www.stat.tamu.edu/~sheather/book/data_sets.php}{Dataset Index Book Webpage}


Class Variable  "Y": "0" or "1"
Attributes used:

1. Length

2. Left

3. Right

4. Bottom
5. Top

6. Diagonal

7. Y
}
\usage{
data(banknote)
}
\description{
Banknote dataset
}
\examples{
data(banknote)

banksmv<-smvcir(group = "Y", data = banknote, test = TRUE)

plotcoords(banksmv, dimensions = 1:3, build_svm = TRUE)

}
\references{
A Modern Approach to Regression with R by Dr. Charles Scheather
}
\keyword{datasets}

