% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_func.R
\name{predict.smvcir}
\alias{predict.smvcir}
\title{Predict new data using SMVCIR}
\usage{
\method{predict}{smvcir}(model, newdata, type = "prob",
  coordinates_only = FALSE, method = "centroid",
  maxdim = model$predictors, kernel = NULL)
}
\arguments{
\item{type}{Either "prob" for class probabilities, or "class" for class predictions.}

\item{coordinates_only}{Default is FALSE.  If TRUE, SMVCIR discriminant coordinates are returned and
can be used towards other classification techniques, such as SVM's or logistic regression.}

\item{maxdim}{Desired number of smvcir dimensions.  A dimensionality test from previously built model
may indicate a sufficient number of dimensions, by default, all possible dimensions are included.}
}
\description{
Takes a set of observations and projects them into the SMVCIR discriminant space based on a previously built SMVCIR model object.
Probabilities are calculated by measuring the mahalonobis distance of an observation from each of the class centroids based on the
original SMVCIR model object.
}

