% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfuncv3.R
\name{smvcir}
\alias{smvcir}
\title{Main SMVCIR function}
\usage{
smvcir(group, data, pdt = 100, level = 0.05, test = FALSE,
  scree_plot = FALSE)
}
\arguments{
\item{group}{A character string specifying the name of the class variable of interest in your dataset.}

\item{data}{A data frame (including your group variable).}

\item{pdt}{Percentage of SMVCIR discrimination desired from dimension}

\item{level}{Level of dimensionality test}

\item{test}{Types of tests, can use either, neither, or both.}

\item{scree_plot}{If TRUE, a scree plot of cumulative percentage of variation explained by discriminant dimensions is produced.}

\item{empsimss}{Number of draws to use in performing each dimensionality test.}
}
\description{
Build a Sliced Mean Variance Covariance Inverse Regression model.
}

